conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
anyio=4.4.0=pyhd8ed1ab_0
asttokens=2.4.1=pyhd8ed1ab_0
bzip2=1.0.8=h99b78c6_7
ca-certificates=2024.8.30=hf0a4a13_0
certifi=2024.8.30=pyhd8ed1ab_0
charset-normalizer=3.3.2=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
decorator=5.1.1=pyhd8ed1ab_0
distro=1.9.0=pyhd8ed1ab_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
filelock=3.16.0=pypi_0
fsspec=2024.9.0=pypi_0
h11=0.14.0=pyhd8ed1ab_0
h2=4.1.0=pyhd8ed1ab_0
hpack=4.0.0=pyh9f0ad1d_0
httpcore=1.0.5=pyhd8ed1ab_0
httpx=0.27.2=pyhd8ed1ab_0
hyperframe=6.0.1=pyhd8ed1ab_0
idna=3.8=pyhd8ed1ab_0
ipython=8.27.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pypi_0
jiter=0.5.0=py312he431725_1
libblas=3.9.0=23_osxarm64_openblas
libcblas=3.9.0=23_osxarm64_openblas
libcxx=18.1.8=h3ed4263_7
libexpat=2.6.3=hf9b8971_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=13_2_0_hd922786_3
libgfortran5=13.2.0=hf226fd6_3
liblapack=3.9.0=23_osxarm64_openblas
libopenblas=0.3.27=openmp_h517c56d_1
libsqlite=3.46.1=hc14010f_0
libzlib=1.3.1=hfb2fe0b_1
llvm-openmp=18.1.8=hde57baf_1
llvmlite=0.43.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
more-itertools=10.5.0=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.5=h7bae524_1
networkx=3.3=pypi_0
numba=0.60.0=pypi_0
numpy=2.0.2=pypi_0
openai=1.44.0=pyhd8ed1ab_0
openai-whisper=20231117=pypi_0
openssl=3.3.2=h8359307_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pip=24.2=pyh8b19718_1
prompt-toolkit=3.0.47=pyha770c72_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pyaudio=0.2.14=pypi_0
pydantic=1.10.17=py312h7e5086c_0
pydub=0.25.1=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
python=3.12.5=h30c5eda_0_cpython
python-dotenv=1.0.1=pypi_0
python_abi=3.12=5_cp312
readline=8.2=h92ec313_1
regex=2024.7.24=pypi_0
requests=2.32.3=pypi_0
scipy=1.14.1=pypi_0
setuptools=73.0.1=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.1=pyhd8ed1ab_0
speechrecognition=3.10.4=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.13.2=pypi_0
tiktoken=0.7.0=pypi_0
tk=8.6.13=h5083fa2_1
torch=2.4.1=pypi_0
tqdm=4.66.5=pyhd8ed1ab_0
traitlets=5.14.3=pyhd8ed1ab_0
typing-extensions=4.12.2=hd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
tzdata=2024a=h8827d51_1
urllib3=2.2.2=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.44.0=pyhd8ed1ab_0
whisper=1.1.10=pypi_0
xz=5.2.6=h57fd34a_0
